rasdaemon: optimization in common code and add support for ACPI RAS2 based memory scrub control
Add optimization of common code and support for ACPI RAS2 based
memory scrub control.

Default value (ACPI RAS on-demand) would be used for default
memory scrub type parameter RAS_DEFAULT_MEM_SCRUB_TYPE if not set.

Default values would be used for RAS2 scrub control
parameters RAS2_CE_THRESHOLD and RAS2_SCRUB_CYCLE_IN_HOUR if not set.

Note: Not sure how to get the memory range(address & size) and
type of scrubber to be used corresponding to a faulty page/address
other than using the deafult memory scrub type set in a system.
Does the scrub control code require to communicate to the scrub
drivers in the kernel, if supported, to findout the right scrubber,
memory range to scrub for the memory/memory range with the faulty page?
If so, the kernel EDAC scrub control need to extend the proposed
sysfs interface for such communication.

Signed-off-by: Shiju Jose <shiju.jose@huawei.com>